[Trinity] Core/Entities: Changed object lowguid to uint64

+++ src/server/game/AI/SmartScripts/SmartScriptMgr.cpp
@@ -130,7 +130,7 @@ void SmartAIMgr::LoadSmartAIFromDB()
 
         SmartScriptHolder temp;
 
-        temp.entryOrGuid = fields[0].GetInt32();
+        temp.entryOrGuid = fields[0].GetInt64();
         SmartScriptType source_type = (SmartScriptType)fields[1].GetUInt8();
         if (source_type >= SMART_SCRIPT_TYPE_MAX)
         {
